+++ packages/fighting-design/_hooks/use-alert-list/interface.d.ts
@@ -1,0 +1,6 @@
+/**
+ * useAlertList 返回值类型接口
+ */
+export interface UseAlertListReturn {
+  startMove: (duration?: number) => void
+}