+++ packages/fighting-design/alert/src/alert.vue
@@ -1,9 +1,9 @@
 <script lang="ts" setup name="FAlert">
   import { Props } from './props'
-  import { ref, reactive } from 'vue'
+  import { ref, reactive, onMounted } from 'vue'
   import { FCloseBtn } from '../../close-btn'
   import { FSvgIcon } from '../../svg-icon'
-  import { useList, useRun, useGlobal } from '../../_hooks'
+  import { useList, useRun, useGlobal, useAlertList } from '../../_hooks'
 
   const prop = defineProps(Props)
 
@@ -25,6 +25,11 @@
   const isShow = ref<boolean>(true)
 
   /**
+   * alertList 节点列表容器 dom
+   */
+  const alertListDom = ref<HTMLElement>(null as unknown as HTMLElement)
+
+  /**
    * 类名列表
    */
   const classList = classes(['type', 'bold', 'simple', 'center', 'round', 'fixed'], 'f-alert')
@@ -43,6 +48,13 @@
     isShow.value = false
     useRun(prop.onClose, evt)
   }
+
+  onMounted((): void => {
+    if (prop.alertList && prop.alertList.length) {
+      const { startMove } = useAlertList(alertListDom.value)
+      startMove()
+    }
+  })
 </script>
 
 <template>
@@ -54,14 +66,7 @@
       </f-svg-icon>
 
       <!-- 主要内容 -->
-      <div
-        :class="[
-          'f-alert__content',
-          {
-            [`f-alert__content-${overflow}`]: overflow
-          }
-        ]"
-      >
+      <div class="f-alert__content">
         <!-- 标题 -->
         <div v-if="title" class="f-alert__title">
           <slot name="title">{{ title }}</slot>
@@ -70,6 +75,11 @@
         <!-- 内容 -->
         <div v-if="$slots.default" class="f-alert__sub-title">
           <slot />
+        </div>
+
+        <!-- 滚动列表 -->
+        <div v-if="alertList && alertList.length" ref="alertListDom" class="f-alert__list">
+          <div v-for="(item, index) in alertList" :key="index" class="f-alert__list-item">{{ item }}</div>
         </div>
       </div>