+++ packages/fighting-design/alert/src/interface.d.ts
@@ -1,6 +1,1 @@
 export type { AlertProps } from './props'
-
-/**
- * 超出展示状态
- */
-export type AlertOverflow = 'hidden'